Create more tests
