+++ libretro-common/compat/compat_fnmatch.c
@@ -20,9 +20,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#if __TEST_FNMATCH__
-#include <assert.h>
-#endif
 #include <stddef.h>
 
 #include <compat/fnmatch.h>
@@ -123,37 +120,3 @@ int rl_fnmatch(const char *pattern, const char *string, int flags)
       return 0;
    return FNM_NOMATCH;
 }
-
-#if __TEST_FNMATCH__
-int main(void)
-{
-   assert(rl_fnmatch("TEST", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE?T", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE[Ssa]T", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE[Ssda]T", "TEsT", 0) == 0);
-   assert(rl_fnmatch("TE[Ssda]T", "TEdT", 0) == 0);
-   assert(rl_fnmatch("TE[Ssda]T", "TEaT", 0) == 0);
-   assert(rl_fnmatch("TEST*", "TEST", 0) == 0);
-   assert(rl_fnmatch("TEST**", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE*ST*", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE**ST*", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE**ST*", "TExST", 0) == 0);
-   assert(rl_fnmatch("TE**ST", "TEST", 0) == 0);
-   assert(rl_fnmatch("TE**ST", "TExST", 0) == 0);
-   assert(rl_fnmatch("TE\\**ST", "TE*xST", 0) == 0);
-   assert(rl_fnmatch("*.*", "test.jpg", 0) == 0);
-   assert(rl_fnmatch("*.jpg", "test.jpg", 0) == 0);
-   assert(rl_fnmatch("*.[Jj][Pp][Gg]", "test.jPg", 0) == 0);
-   assert(rl_fnmatch("*.[Jj]*[Gg]", "test.jPg", 0) == 0);
-   assert(rl_fnmatch("TEST?", "TEST", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TES[asd", "TEST", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TEST\\", "TEST", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TEST*S", "TEST", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TE**ST", "TExT", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TE\\*T", "TExT", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TES?", "TES", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TE", "TEST", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("TEST!", "TEST", 0) == FNM_NOMATCH);
-   assert(rl_fnmatch("DSAD", "TEST", 0) == FNM_NOMATCH);
-}
-#endif
